+++ src/packages/radiogroup/index.js
@@ -1,4 +1,5 @@
 import RadioGroup from './radiogroup.vue';
+import './radiogroup.scss';
 
 RadioGroup.install = function(Vue) {
   Vue.component(RadioGroup.name, RadioGroup);